migrated CacheMixin to a separated file
