+++ thegame/views.py
@@ -5,7 +5,6 @@ from django.shortcuts import redirect
 from django.urls import reverse
 from django.views import View
 from django.views.generic import TemplateView
-from django.views.decorators.cache import cache_page
 
 
 class FilesView(LoginRequiredMixin, View):
@@ -25,17 +24,3 @@ class IndexView(TemplateView):
         if request.user.username is None:
             return redirect(reverse('set_username'))
         return super().dispatch(request, *args, **kwargs)
-
-
-class CacheMixin(object):
-    cache_timeout = 60
-
-    def get_cache_timeout(self):
-        force_update = self.request.GET.get('force_update', None)
-        if force_update == 'true':
-            # Disable cache if the query param: force_update=true is recieved.
-            return 0
-        return self.cache_timeout
-
-    def dispatch(self, *args, **kwargs):
-        return cache_page(self.get_cache_timeout())(super(CacheMixin, self).dispatch)(*args, **kwargs)
